% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{getSkeletalInventories}
\alias{getSkeletalInventories}
\title{get all skeleletal inventory labels}
\usage{
getSkeletalInventories(endpoint = "http://localhost/ds/query")
}
\arguments{
\item{endpoint}{url specifying sparql endpoint}
}
\description{
get all skeleletal inventory labels
}
