% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.r
\name{getSkeletalInventoryLabels}
\alias{getSkeletalInventoryLabels}
\title{get all skeleletal inventory labels}
\usage{
getSkeletalInventoryLabels(endpoint = "http://localhost:2020/ds/query")
}
\arguments{
\item{endpoint}{url specifying sparql endpoint}
}
\description{
get all skeleletal inventory labels
}
