% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVariables.r
\docType{data}
\name{prefixes}
\alias{prefixes}
\alias{RDFBonesPrefix}
\alias{RDFBonesPrefixString}
\title{RDF Prefixes needed to perform SPARQL queries on RDFBones installation}
\format{An object of class \code{character} of length 56.}
\usage{
RDFBonesPrefix

RDFBonesPrefixString
}
\description{
RDF Prefixes needed to perform SPARQL queries on RDFBones installation
}
\keyword{datasets}
