% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDFBonesQuery-package.r
\docType{package}
\encoding{utf8}
\name{RDFBonesQuery-package}
\alias{RDFBonesQuery-package}
\alias{RDFBonesQuery}
\title{Provide predefined queries of RDFBones SPARQL endpoint}
\description{
Provide predefined queries of RDFBones SPARQL endpoint
}
\details{
\tabular{ll}{
Package: \tab RDFBonesQuery\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2017-03-16\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\note{
The pdf-version of Morpho-help can be obtained from CRAN on \url{https://cran.r-project.org/package=Morpho}

For more advanced operations on triangular surface meshes, check out my package Rvcg: \url{https://cran.r-project.org/package=Rvcg} or the code repository on github \url{https://github.com/zarquon42b/Rvcg}
}
\references{
Schlager S. 2013. Soft-tissue reconstruction of the human nose:
population differences and sexual dimorphism. PhD thesis,
\enc{Universitätsbibliothek}{Universitaetsbibliothek} Freiburg. URL:
\url{http://www.freidok.uni-freiburg.de/volltexte/9181/}.
}
\author{
Stefan Schlager \email{stefan.schlager@uniklinik-freiburg.de}

Maintainer: Stefan Schlager \email{stefan.schlager@uniklinik-freiburg.de}
}
\keyword{package}
