% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDFBonesQuery-package.r
\docType{data}
\name{RDFBonesVariables}
\alias{RDFBonesVariables}
\alias{RDFBonesPrefix}
\alias{skull_0144_ch_fe.mesh}
\title{Landmarks and a triangular mesh}
\format{\code{RDFBonesPrefix}: character vector containing namespace prefix definitions.

\code{RDFBonesPrefixString}: character specifying all NS-Prefixes for SPARQL queries.}
\description{
Landmarks on the osseous human nose and a triangular mesh representing this
structure.
}
\keyword{datasets}
